feat: Add AlbumSeeder to seed a default album
Add a new seeder class, AlbumSeeder, to seed a default album into the database. The album has the title "Greatest Hits" and is associated with the artist "Blessd". It has a release date of September 1, 2024, and a description of "The best songs of all time." The album's icon is set to a specific URL.