feat: Add PostSeeder to seed a default post
This commit adds the PostSeeder class to the seeders directory. The PostSeeder is responsible for creating a new post in the database. It sets the user_id, content, image, and song_id fields of the post. The user_id is hardcoded to a specific user, the content is set to 'This is my favorite song', the image is set to a specific URL, and the song_id is retrieved based on the artist_id of a specific artist. This seeder is used to populate the database with a default post.