feat: Add FollowedPlaylistSeeder to seed followed playlists
Add FollowedPlaylistSeeder to the DatabaseSeeder to seed followed playlists for users. This will allow users to have pre-defined playlists that they can follow and listen to.

Also, update the UserSeeder to create two additional users: Valentina and Admin. Valentina is a regular user with a profile picture and banner, while Admin is an administrator with additional privileges.