Add Facebook auth and basedata changes
This commit adds a way to login with a facebook account. It implements GET routes on auth.php that redirect to main functions on authcontroller.php.

Some changes on basedata were added such as:
-make the password field as "nullable"
-create a facebook_id field
-update username field collation to utf8_bin

The purpoise of these changes on basedata is that google auth and others needs a password to be nullable because when you are loggin or create it with, you don't need a defined password because of your third service account count as a key. The facebook_id was added beacuase of login needs to remember and verify  this personal id.  And to conclude, the reason for username collation is that change allows to be case-sensitive, fixing the problem with duplicate usernames.