fix: Login now requires 2fa code
Now, if the user requires to login on the app and has the 2fa enabled, the login form will return this: "'Two-factor authentication code required'], 400);". Then for login, the user has to go to google authenticator app and copy the 2fa code and paste it on the request called: "verify-secret", then now the request will validate if that code is correct and if it is, will show a message 200 validating the sussesful login and the login token.

If the user has not enabled 2fa, login request will aloow it to login on the app without any problem